rearrange LP stuff into their functions
