+++ src/model.h
@@ -60,101 +60,10 @@ typedef struct {
 } message;
 
 /*******************************************/
-/**************** Station LP ***************/
-
-// Station state
-
-typedef enum {
-    ST_EMPTY, // Station is empty, let trains in
-    ST_OCCUPIED, // There is a train here, alighting passengers
-    ST_BOARDING, // Loading passengers onto the TU
-} station_sm; //SM - state machine, not full state
-
-// Wraps up the track state
-typedef struct {
-    station_sm inbound;
-    station_sm outbound;
-    
-    int track_id; // Just an identifier so we can tell them apart
-
-    tw_lpid curr_tu; // who is on the track now
-
-    // Inbound queue
-    unsigned short queued_tu_present;
-    tw_lpid queued_tu[QUEUE_LEN];
-
-    // Outbound queue
-    // This really represents the space between the stations`
-    int next_arrival;
-
-    short from_queue;
-
-} track_t;
-
-typedef struct {
-    char station_name[25]; // Identifier name for station TODO FIX SIZE
-
-    //station_sm curr_state; // What state is it in right now?
-
-    // The left track is where the station ID of the adjacent station is less
-    // The righttrack is where the station ID of the adjacent station is less
-    track_t left;
-    track_t right;
-
-    //unsigned short queued_tu_present; // Anything in the queue?
-    //tw_lpid queued_tu; // The TU queued up. TODO: For now a single int, should get expanded
-
-    passenger_t* pass_list; //Passenger linked list
-} station_state;
-
-//Global variables used by both main and driver
-// - this defines the LP types
-extern tw_lptype station_lps[];
-
-//Function Declarations
-// defined in station.c:
-extern void station_init(station_state *s, tw_lp *lp);
-extern void station_event(station_state *s, tw_bf *bf, message *in_msg, tw_lp *lp);
-extern void station_event_reverse(station_state *s, tw_bf *bf, message *in_msg, tw_lp *lp);
-extern void station_final(station_state *s, tw_lp *lp);
 // defined in model_map.c:
 extern tw_peid station_map(tw_lpid gid);
 
 /*******************************************/
-/************* Transit Unit LP *************/
-
-typedef enum {
-    TU_IDLE, // Starting state only
-    TU_APPROACH, // Approaching a station
-    TU_ALIGHT, // At a station, empty pass
-    TU_BOARD, // At a station, barding pass
-} transit_unit_sm; //sm - state machine, diff from state
-
-// TU state
-typedef struct {
-    transit_unit_sm curr_state; // Some kind of state machine?
-    int start; // time to start
-    tw_lpid prev_station; // GID of previous station
-    tw_lpid station; // GID of current station
-    int route_index; // Current index in route
-    // This should be an array of some kind
-    struct route_t* route; 
-    // TODO Almost certainly we want to sort these by where they are getting off...
-    // for now, a big stupid linked list
-    passenger_t* pass_list; //Passenger linked list
-    int pass_count;
-
-    int min_time; // the minimum arrival time at next station
-
-} tu_state;
-
-//Function Declarations
-// defined in transit_unit.c:
-extern void transit_unit_init(tu_state *s, tw_lp *lp);
-extern void transit_unit_pre_run(tu_state *s, tw_lp *lp);
-extern void transit_unit_event(tu_state *s, tw_bf *bf, message *in_msg, tw_lp *lp);
-extern void transit_unit_event_reverse(tu_state *s, tw_bf *bf, message *in_msg, tw_lp *lp);
-extern void transit_unit_final(tu_state *s, tw_lp *lp);
 
 extern tw_peid transit_unit_map(tw_lpid gid);
