+++ src/station.c
@@ -10,6 +10,7 @@
 #include "passenger.h"
 #include "model.h"
 #include "graph.h"
+#include "station.h"
 #include "track.h"
 
 //Helper Functions
